+++ src/universe.h
@@ -5,23 +5,23 @@
 class Universe
 {
 public:
-    double dt;
-    double repulsion;
-    double spring_k;
-    double damping;
-    double gravity;
+    float dt;
+    float repulsion;
+    float spring_k;
+    float damping;
+    float gravity;
     int n_iterations;
     Graph graph;
     Universe(Graph graph,
-        double dt,
-        double repulsion,
-        double spring_k,
-        double damping,
-        double gravity);
+             float dt,
+             float repulsion,
+             float spring_k,
+             float damping,
+             float gravity);
 
-    void update(double deltaT);
+    void update(float deltaT);
     Vec3D compute_spring_force(Node n1, Node n2);
     Vec3D compute_repulsion_force(Node n1, Node n2);
-    Vec3D compute_spring_force_general(double k, Vec3D v1, Vec3D v2);
+    Vec3D compute_spring_force_general(float k, Vec3D v1, Vec3D v2);
     void set_graph(Graph graph);
 };
